Revert "First working built-in http proxy"
This reverts commit 5ba0dffa5d4eaa180dd8d7015d7f76cd0c4be463.
