+++ shadowsocks-csharp/Controller/ShadowsocksController.cs
@@ -8,7 +8,7 @@ using System.Threading;
 using System.Web;
 using System.Windows.Forms;
 using Newtonsoft.Json;
-using Shadowsocks.Controller.Service;
+
 using Shadowsocks.Controller.Strategy;
 using Shadowsocks.Model;
 using Shadowsocks.Properties;
@@ -463,7 +463,6 @@ namespace Shadowsocks.Controller
                 services.Add(tcpRelay);
                 services.Add(udpRelay);
                 services.Add(_pacServer);
-                services.Add(new Http2Socks5(_config.localPort));
                 _listener = new Listener(services);
                 _listener.Start(_config);
             }
